requirements.txt
Pillow
ips.py